+++ ulmo/runs/PowerLaw/modis_powerlaw.yaml
@@ -1,0 +1,63 @@
+# Script to perform VIIRS 2020 Pre-proc
+apiVersion: batch/v1
+kind: Job
+metadata:
+  name: xavier-modis-powerlaw-debug
+spec:
+  backoffLimit: 0
+  template:
+    spec:
+      containers:
+      - name: container
+        image: localhost:30081/profxj/ulmo_nvidia:latest  # UPDATE
+        imagePullPolicy: Always
+        resources:
+          requests:
+            cpu: "1"
+            memory: "16Gi"  # Took 60Gi+ on my machine
+            ephemeral-storage: 30Gi  # Fine if compressed
+          limits:
+            cpu: "2"
+            memory: "30Gi"
+            ephemeral-storage: 200Gi
+            #nvidia.com/gpu:  "1"  # See docs to exlude certain types
+        command: ["/bin/bash", "-c"]
+        args:
+          - cd ulmo; 
+            git fetch;
+            git checkout modis_ssl_full; 
+            git pull; 
+            python setup.py develop;
+            cd ../tmp; 
+            mkdir foo; 
+            cd foo; 
+            python -u /ulmo/ulmo/runs/PowerLaw/modis_powerlaw.py --task powerlaw --debug;
+        env:
+          - name: "ENDPOINT_URL"
+            value: "http://rook-ceph-rgw-nautiluss3.rook"
+          - name: "S3_ENDPOINT"
+            value: "rook-ceph-rgw-nautiluss3.rook"
+        volumeMounts:
+          - name: prp-s3-credentials
+            mountPath: "/root/.aws/credentials"
+            subPath: "credentials"
+          - name: ephemeral
+            mountPath: "/tmp"
+          - name: "dshm"
+            mountPath: "/dev/shm"
+      nodeSelector:
+        nautilus.io/disktype: nvme
+        gpu-type: "1080Ti"  # titan-xp
+      restartPolicy: Never
+      volumes:
+        # Secrets file for nautilus s3 credentials .aws/credentials and .s3cfg
+        - name: prp-s3-credentials
+          secret:
+            secretName: prp-s3-credentials
+        # Shared memory (necessary for Python's multiprocessing.shared_memory module to work)
+        - name: dshm
+          emptyDir:
+            medium: Memory
+        # Ephemeral storage
+        - name: ephemeral
+          emptyDir: {}
